+++ Tool/Sources/OpenAIService/Memory/AutoManagedChatGPTMemory.swift
@@ -1,77 +1,6 @@
 import Foundation
 import GPTEncoder
-
-public protocol ChatGPTMemory {
-    /// The visible messages to the ChatGPT service.
-    var messages: [ChatMessage] { get async }
-    /// The remaining tokens available for the reply.
-    var remainingTokens: Int? { get async }
-    /// Update the message history.
-    func mutateHistory(_ update: (inout [ChatMessage]) -> Void) async
-}
-
-public extension ChatGPTMemory {
-    /// Append a message to the history.
-    func appendMessage(_ message: ChatMessage) async {
-        await mutateHistory {
-            $0.append(message)
-        }
-    }
-
-    /// Update a message in the history.
-    func updateMessage(id: String, _ update: (inout ChatMessage) -> Void) async {
-        await mutateHistory { history in
-            if let index = history.firstIndex(where: { $0.id == id }) {
-                update(&history[index])
-            }
-        }
-    }
-
-    /// Remove a message from the history.
-    func removeMessage(_ id: String) async {
-        await mutateHistory {
-            $0.removeAll { $0.id == id }
-        }
-    }
-
-    /// Stream a message to the history.
-    func streamMessage(id: String, role: ChatMessage.Role?, content: String?) async {
-        await mutateHistory { history in
-            if let index = history.firstIndex(where: { $0.id == id }) {
-                if let content {
-                    history[index].content.append(content)
-                }
-                if let role {
-                    history[index].role = role
-                }
-            } else {
-                history.append(.init(
-                    id: id,
-                    role: role ?? .system,
-                    content: content ?? ""
-                ))
-            }
-        }
-    }
-
-    /// Clear the history.
-    func clearHistory() async {
-        await mutateHistory { $0.removeAll() }
-    }
-}
-
-public actor ConversationChatGPTMemory: ChatGPTMemory {
-    public var messages: [ChatMessage] = []
-    public var remainingTokens: Int? { nil }
-
-    public init(systemPrompt: String) {
-        messages.append(.init(role: .system, content: systemPrompt))
-    }
-
-    public func mutateHistory(_ update: (inout [ChatMessage]) -> Void) {
-        update(&messages)
-    }
-}
+import Preferences
 
 /// A memory that automatically manages the history according to max tokens and max message count.
 public actor AutoManagedChatGPTMemory: ChatGPTMemory {
@@ -100,6 +29,13 @@ public actor AutoManagedChatGPTMemory: ChatGPTMemory {
 
     public func mutateSystemPrompt(_ newPrompt: String) {
         systemPrompt.content = newPrompt
+    }
+    
+    public nonisolated
+    func observeHistoryChange(_ onChange: @escaping () -> Void) {
+        Task {
+            await setOnHistoryChangeBlock(onChange)
+        }
     }
 
     func generateSendingHistory(
@@ -144,13 +80,6 @@ public actor AutoManagedChatGPTMemory: ChatGPTMemory {
         )
         .reduce(0) { $0 + ($1.tokensCount ?? 0) }
         return max(configuration.minimumReplyTokens, configuration.maxTokens - tokensCount)
-    }
-
-    public nonisolated
-    func observeHistoryChange(_ onChange: @escaping () -> Void) {
-        Task {
-            await setOnHistoryChangeBlock(onChange)
-        }
     }
 
     func setOnHistoryChangeBlock(_ onChange: @escaping () -> Void) {
